+++ spec/lib/aixm/schedule/date_spec.rb
@@ -73,33 +73,33 @@ describe AIXM::Schedule::Date do
       end
     end
 
-    describe :prev do
+    describe :pred do
       it "returns a new object of the preceding day" do
         date = AIXM.date('2000-06-07')
-        subject = date.prev
+        subject = date.pred
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('2000-06-06')
       end
 
       it "returns a new object of the preceding day across year boundaries" do
         date = AIXM.date('2001-01-01')
-        subject = date.prev
+        subject = date.pred
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('2000-12-31')
       end
     end
 
-    describe :next do
+    describe :succ do
       it "returns a new object of the following day" do
         date = AIXM.date('2000-06-07')
-        subject = date.next
+        subject = date.succ
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('2000-06-08')
       end
 
       it "returns a new object of the following day across year boundaries" do
         date = AIXM.date('2000-12-31')
-        subject = date.next
+        subject = date.succ
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('2001-01-01')
       end
@@ -298,33 +298,33 @@ describe AIXM::Schedule::Date do
       end
     end
 
-    describe :prev do
+    describe :pred do
       it "returns a new object of the preceding day" do
         date = AIXM.date('06-07')
-        subject = date.prev
+        subject = date.pred
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('06-06')
       end
 
       it "returns a new object of the preceding day across year boundaries" do
         date = AIXM.date('01-01')
-        subject = date.prev
+        subject = date.pred
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('12-31')
       end
     end
 
-    describe :next do
+    describe :succ do
       it "returns a new object of the following day" do
         date = AIXM.date('06-07')
-        subject = date.next
+        subject = date.succ
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('06-08')
       end
 
       it "returns a new object of the following day across year boundaries" do
         date = AIXM.date('12-31')
-        subject = date.next
+        subject = date.succ
         _(subject.object_id).wont_equal date.object_id
         _(subject).must_equal AIXM.date('01-01')
       end